refactor(desktop): extract ChatPane mid-tier components (Phase 3/5)
- ChatHeader.tsx
- QueuedSessionInputRail.tsx
- AssistantTurn/{ActionsMenu,Outputs,MemoryProposals,TraceStepGroup}.tsx
- AssistantTurn/index.tsx — orchestrates the cluster
