+++ desktop/src/components/panes/ChatPane.tsx
@@ -217,13 +217,9 @@ import {
   sortOutputsLatestFirst,
   dedupeOutputsForDisplay,
 } from "./ChatPane/ArtifactBrowserModal";
-import {
-  ExecutionTimelineThinkingEntry,
-  LiveStatusEllipsis,
-  LiveStatusLine,
-  StreamingCursor,
-  TraceTimelineStepEntry,
-} from "./ChatPane/AssistantTurn/status";
+import { ChatHeader } from "./ChatPane/ChatHeader";
+import { QueuedSessionInputRail } from "./ChatPane/QueuedSessionInputRail";
+import { AssistantTurn } from "./ChatPane/AssistantTurn";
 
 export type {
   ChatAssistantSegment,
@@ -237,6 +233,7 @@ export {
   turnInputIdsFromHistoryMessages,
   ArtifactBrowserModal,
   sortOutputs,
+  AssistantTurn,
 };
 
 function sessionUserId(
@@ -7624,96 +7621,6 @@ export function ChatPane({
   );
 }
 
-interface ChatHeaderProps {
-  agentName: string;
-  workspace: WorkspaceRecordPayload | null;
-  subtitle?: string;
-  onReturnToMainSession?: () => void;
-  onOpenSessions?: () => void;
-  onOpenInbox?: () => void;
-  inboxUnreadCount: number;
-  onOpenAutomations?: () => void;
-}
-
-function ChatHeader({
-  agentName,
-  workspace,
-  subtitle,
-  onReturnToMainSession,
-  onOpenSessions,
-  onOpenInbox,
-  inboxUnreadCount,
-  onOpenAutomations,
-}: ChatHeaderProps) {
-  const seed = workspace?.id ?? agentName ?? "default";
-
-  return (
-    <div className="flex items-center justify-between gap-3">
-      <div className="flex min-w-0 flex-1 items-center gap-2">
-        <AgentAvatar seed={seed} size="sm" />
-        <div className="flex min-w-0 flex-col">
-          <span className="truncate text-sm font-medium text-foreground">
-            {agentName}
-          </span>
-          {subtitle ? (
-            <span className="truncate text-[11px] leading-tight text-muted-foreground">
-              {subtitle}
-            </span>
-          ) : null}
-        </div>
-      </div>
-
-      <div className="flex shrink-0 items-center gap-0.5">
-        {onOpenSessions ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenSessions()}
-            aria-label="Sessions"
-            className="text-muted-foreground hover:text-foreground"
-          >
-            <MessageCircle className="size-4" />
-          </Button>
-        ) : null}
-
-        {onOpenInbox ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenInbox()}
-            aria-label="Inbox"
-            className="relative text-muted-foreground hover:text-foreground"
-          >
-            <Inbox className="size-4" />
-            {inboxUnreadCount > 0 ? (
-              <StatusDot
-                variant="destructive"
-                size="sm"
-                className="absolute right-1 top-1 border border-card"
-              />
-            ) : null}
-          </Button>
-        ) : null}
-
-        {onOpenAutomations ? (
-          <Button
-            type="button"
-            variant="ghost"
-            size="icon-sm"
-            onClick={() => onOpenAutomations()}
-            aria-label="Automations"
-            className="text-muted-foreground hover:text-foreground"
-          >
-            <Clock3 className="size-4" />
-          </Button>
-        ) : null}
-      </div>
-    </div>
-  );
-}
-
 interface ComposerProps {
   input: string;
   quotedSkills: ChatComposerQuotedSkillItem[];
@@ -7933,798 +7840,6 @@ function UserTurnComponent({
   );
 }
 
-function queuedSessionInputPreviewText(item: QueuedSessionInput) {
-  const parsedQuotedSkills = parseSerializedQuotedSkillPrompt(item.text);
-  const previewText =
-    parsedQuotedSkills.body ||
-    parsedQuotedSkills.skillIds.map((skillId) => `/${skillId}`).join(" ");
-  return previewText.replace(/\s+/g, " ").trim();
-}
-
-function QueuedSessionInputRail({
-  items,
-  onEditItem,
-  children,
-}: {
-  items: QueuedSessionInput[];
-  onEditItem?: (item: QueuedSessionInput, nextText: string) => Promise<void>;
-  children: ReactNode;
-}) {
-  const [editingInputId, setEditingInputId] = useState("");
-  const [editingDraft, setEditingDraft] = useState("");
-  const [editingError, setEditingError] = useState("");
-  const [savingInputId, setSavingInputId] = useState("");
-  const panelInsetPx = 18;
-  const panelHeightPx = 112;
-  const overlapPx = 28;
-  const queueViewportHeightPx = 50;
-  const reservedTopPx = 94;
-
-  useEffect(() => {
-    if (!editingInputId) {
-      return;
-    }
-    const activeItem = items.find((item) => item.inputId === editingInputId);
-    if (!activeItem || activeItem.status !== "queued") {
-      setEditingInputId("");
-      setEditingDraft("");
-      setEditingError("");
-      setSavingInputId("");
-    }
-  }, [editingInputId, items]);
-
-  const cancelEditing = () => {
-    setEditingInputId("");
-    setEditingDraft("");
-    setEditingError("");
-    setSavingInputId("");
-  };
-
-  const saveEditingItem = async (item: QueuedSessionInput) => {
-    if (!onEditItem || savingInputId || item.status !== "queued") {
-      return;
-    }
-    setEditingError("");
-    setSavingInputId(item.inputId);
-    try {
-      await onEditItem(item, editingDraft);
-      cancelEditing();
-    } catch (error) {
-      setEditingError(normalizeErrorMessage(error));
-    } finally {
-      setSavingInputId("");
-    }
-  };
-
-  if (items.length === 0) {
-    return <>{children}</>;
-  }
-
-  return (
-    <div className="relative" style={{ paddingTop: `${reservedTopPx}px` }}>
-      <div className="pointer-events-none absolute inset-x-0 top-0">
-        <div
-          className="pointer-events-auto absolute inset-x-0 overflow-hidden rounded-3xl bg-background shadow-md"
-          style={{
-            left: `${panelInsetPx}px`,
-            right: `${panelInsetPx}px`,
-            height: `${panelHeightPx}px`,
-          }}
-        >
-          <div className="px-4 pt-4">
-            <div
-              className="overflow-y-auto pr-1.5"
-              style={{ maxHeight: `${queueViewportHeightPx}px` }}
-            >
-              <div className="space-y-1.5">
-                {items.map((item) => {
-                  const previewText = queuedSessionInputPreviewText(item);
-                  const isEditing = editingInputId === item.inputId;
-                  const isSaving = savingInputId === item.inputId;
-                  return (
-                    <div
-                      key={item.inputId}
-                      className="rounded-xl px-1 text-sm leading-7 text-foreground"
-                    >
-                      {isEditing ? (
-                        <div className="space-y-1.5">
-                          <div className="flex items-center gap-2">
-                            <CornerDownLeft className="size-4 shrink-0 text-muted-foreground" />
-                            <Input
-                              value={editingDraft}
-                              onChange={(event) =>
-                                setEditingDraft(event.target.value)
-                              }
-                              onKeyDown={(event) => {
-                                if (event.key === "Enter") {
-                                  event.preventDefault();
-                                  void saveEditingItem(item);
-                                } else if (event.key === "Escape") {
-                                  event.preventDefault();
-                                  cancelEditing();
-                                }
-                              }}
-                              disabled={isSaving}
-                              autoFocus
-                              className="h-8 min-w-0 flex-1 rounded-lg border-border bg-background px-2.5 text-sm"
-                            />
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              disabled={isSaving}
-                              onClick={() => {
-                                void saveEditingItem(item);
-                              }}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Save queued message edit"
-                            >
-                              {isSaving ? (
-                                <Loader2 className="size-3.5 animate-spin" />
-                              ) : (
-                                <Check className="size-3.5" />
-                              )}
-                            </Button>
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              disabled={isSaving}
-                              onClick={cancelEditing}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Cancel queued message edit"
-                            >
-                              <X className="size-3.5" />
-                            </Button>
-                          </div>
-                          {editingError ? (
-                            <div className="pl-6 text-xs leading-5 text-destructive">
-                              {editingError}
-                            </div>
-                          ) : null}
-                        </div>
-                      ) : (
-                        <div className="flex items-center gap-3">
-                          <CornerDownLeft className="size-4 shrink-0 text-muted-foreground" />
-                          <div className="min-w-0 flex-1 truncate">
-                            {previewText || "Queued message"}
-                          </div>
-                          {onEditItem && item.status === "queued" ? (
-                            <Button
-                              type="button"
-                              variant="ghost"
-                              size="icon-xs"
-                              onClick={() => {
-                                setEditingInputId(item.inputId);
-                                setEditingDraft(previewText);
-                                setEditingError("");
-                              }}
-                              className="size-7 rounded-full text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                              aria-label="Edit queued message"
-                            >
-                              <PencilLine className="size-3.5" />
-                            </Button>
-                          ) : null}
-                        </div>
-                      )}
-                    </div>
-                  );
-                })}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div
-        className="relative z-10 rounded-3xl bg-background"
-        style={{ marginTop: `${-overlapPx}px` }}
-      >
-        {children}
-      </div>
-    </div>
-  );
-}
-
-function executionItemsHaveFileEdits(
-  items: ChatExecutionTimelineItem[],
-): boolean {
-  if (items.length === 0) {
-    return false;
-  }
-  return items.some((item) => {
-    if (item.kind !== "trace_step" || item.step.kind !== "tool") {
-      return false;
-    }
-    const title = item.step.title.toLowerCase();
-    return (
-      title.startsWith("edit") ||
-      title.startsWith("write") ||
-      title.startsWith("patch") ||
-      title.startsWith("replace") ||
-      title.startsWith("multiedit") ||
-      title.startsWith("apply") ||
-      title.startsWith("create file")
-    );
-  });
-}
-
-interface AssistantTurnActionsMenuProps {
-  copyText: string;
-  onViewTurnDetails?: () => void;
-  onViewFileChanges?: () => void;
-  hasFileEdits?: boolean;
-}
-
-function AssistantTurnActionsMenu({
-  copyText,
-  onViewTurnDetails,
-  onViewFileChanges,
-  hasFileEdits,
-}: AssistantTurnActionsMenuProps) {
-  const [copied, setCopied] = useState(false);
-
-  async function handleCopy() {
-    if (!copyText) {
-      return;
-    }
-    try {
-      await navigator.clipboard.writeText(copyText);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 1500);
-    } catch {
-      // ignore
-    }
-  }
-
-  const showFileChanges = Boolean(onViewFileChanges) && Boolean(hasFileEdits);
-  const canCopy = copyText.trim().length > 0;
-  const hasOverflowItems =
-    Boolean(onViewTurnDetails) || showFileChanges;
-
-  if (!canCopy && !hasOverflowItems) {
-    return null;
-  }
-
-  return (
-    <div className="flex items-center gap-0.5">
-      {canCopy ? (
-        <Button
-          aria-label={copied ? "Copied message" : "Copy message"}
-          className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-          size="icon-xs"
-          type="button"
-          variant="ghost"
-          onClick={() => void handleCopy()}
-        >
-          {copied ? (
-            <Check className="size-3.5" strokeWidth={1.9} />
-          ) : (
-            <Copy className="size-3.5" strokeWidth={1.9} />
-          )}
-        </Button>
-      ) : null}
-      {hasOverflowItems ? (
-        <DropdownMenu>
-          <DropdownMenuTrigger
-            render={
-              <Button
-                aria-label="Turn actions"
-                className="size-6 rounded-lg text-muted-foreground hover:bg-fg-6 hover:text-foreground"
-                size="icon-xs"
-                type="button"
-                variant="ghost"
-              >
-                <MoreHorizontal className="size-3.5" strokeWidth={1.9} />
-              </Button>
-            }
-          />
-          <DropdownMenuContent align="end" className="w-44" sideOffset={4}>
-            {onViewTurnDetails ? (
-              <DropdownMenuItem onClick={onViewTurnDetails}>
-                <ListTree />
-                View turn details
-              </DropdownMenuItem>
-            ) : null}
-            {showFileChanges ? (
-              <DropdownMenuItem onClick={onViewFileChanges}>
-                <FileCode2 />
-                View file changes
-              </DropdownMenuItem>
-            ) : null}
-          </DropdownMenuContent>
-        </DropdownMenu>
-      ) : null}
-    </div>
-  );
-}
-
-export const AssistantTurn = memo(AssistantTurnComponent, (prev, next) =>
-  prev.label === next.label &&
-  prev.mode === next.mode &&
-  prev.showExecutionInternals === next.showExecutionInternals &&
-  prev.text === next.text &&
-  prev.tone === next.tone &&
-  prev.segments === next.segments &&
-  prev.executionItems === next.executionItems &&
-  prev.memoryProposals === next.memoryProposals &&
-  prev.outputs === next.outputs &&
-  prev.memoryProposalAction === next.memoryProposalAction &&
-  prev.editingMemoryProposalId === next.editingMemoryProposalId &&
-  prev.memoryProposalDrafts === next.memoryProposalDrafts &&
-  prev.collapsedTraceByStepId === next.collapsedTraceByStepId &&
-  prev.live === next.live &&
-  prev.status === next.status &&
-  prev.statusAccessory === next.statusAccessory &&
-  prev.footerAccessory === next.footerAccessory,
-);
-
-function AssistantTurnComponent({
-  label,
-  mode,
-  showExecutionInternals = true,
-  fitToContent = false,
-  text,
-  tone = "default",
-  segments,
-  executionItems,
-  memoryProposals,
-  outputs,
-  memoryProposalAction,
-  editingMemoryProposalId,
-  memoryProposalDrafts,
-  onEditMemoryProposal,
-  onMemoryProposalDraftChange,
-  onAcceptMemoryProposal,
-  onDismissMemoryProposal,
-  onOpenOutput,
-  onOpenAllArtifacts,
-  collapsedTraceByStepId,
-  onToggleTraceStep,
-  onLinkClick,
-  onLocalLinkClick,
-  showAvatar = false,
-  workspaceId = null,
-  createdAt,
-  status = "",
-  live = false,
-  statusAccessory = null,
-  footerAccessory = null,
-}: {
-  label: string;
-  mode: string;
-  showExecutionInternals?: boolean;
-  fitToContent?: boolean;
-  text: string;
-  tone?: ChatMessage["tone"];
-  segments: ChatAssistantSegment[];
-  executionItems: ChatExecutionTimelineItem[];
-  memoryProposals: MemoryUpdateProposalRecordPayload[];
-  outputs: WorkspaceOutputRecordPayload[];
-  memoryProposalAction: {
-    proposalId: string;
-    action: "accept" | "dismiss";
-  } | null;
-  editingMemoryProposalId: string | null;
-  memoryProposalDrafts: Record<string, string>;
-  onEditMemoryProposal: (proposalId: string) => void;
-  onMemoryProposalDraftChange: (proposalId: string, value: string) => void;
-  onAcceptMemoryProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissMemoryProposal: (
-    proposal: MemoryUpdateProposalRecordPayload,
-  ) => void;
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-  collapsedTraceByStepId: Record<string, boolean>;
-  onToggleTraceStep: (stepId: string) => void;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  showAvatar?: boolean;
-  workspaceId?: string | null;
-  createdAt?: string;
-  status?: string;
-  live?: boolean;
-  statusAccessory?: ReactNode;
-  footerAccessory?: ReactNode;
-}) {
-  const normalizedStatus = (
-    showExecutionInternals ? status : status ? "Working" : ""
-  )
-    .replace(/\.+$/, "")
-    .trim();
-  const visibleSegments = showExecutionInternals
-    ? segments
-    : segments.filter(
-        (segment): segment is Extract<ChatAssistantSegment, { kind: "output" }> =>
-          segment.kind === "output",
-      );
-  const visibleExecutionItems = showExecutionInternals ? executionItems : [];
-  const renderedSegments =
-    visibleSegments.length > 0
-      ? visibleSegments
-      : visibleExecutionItems.length > 0 || Boolean(text)
-        ? [
-            ...(visibleExecutionItems.length > 0
-              ? ([
-                  {
-                    kind: "execution",
-                    items: visibleExecutionItems,
-                  },
-                ] as ChatAssistantSegment[])
-              : []),
-            ...(text
-              ? ([
-                  {
-                    kind: "output",
-                    text,
-                    tone,
-                  },
-                ] as ChatAssistantSegment[])
-              : []),
-          ]
-        : [];
-  const showStatusPlaceholder =
-    live && Boolean(normalizedStatus) && renderedSegments.length === 0;
-  const lastSegmentIsOutput =
-    renderedSegments.length > 0 &&
-    renderedSegments[renderedSegments.length - 1]?.kind === "output";
-  const showWorkingStatusLine =
-    live &&
-    showExecutionInternals &&
-    renderedSegments.length > 0 &&
-    !lastSegmentIsOutput;
-  const showStreamingCursor = live && lastSegmentIsOutput;
-
-  const [forceExpandToken, setForceExpandToken] = useState(0);
-  const hasFileEdits = useMemo(
-    () => executionItemsHaveFileEdits(executionItems),
-    [executionItems],
-  );
-  const copyText = useMemo(
-    () =>
-      renderedSegments
-        .filter(
-          (segment): segment is Extract<ChatAssistantSegment, { kind: "output" }> =>
-            segment.kind === "output",
-        )
-        .map((segment) => segment.text)
-        .join("\n\n")
-        .trim() || text.trim(),
-    [renderedSegments, text],
-  );
-  const hasAnyContent = renderedSegments.length > 0;
-  const showActionsMenu = hasAnyContent && !live;
-  const renderStatusLine = (nextLabel: string, className = "") => {
-    const resolvedLabel = nextLabel.trim() || "Working";
-    if (!statusAccessory) {
-      return <LiveStatusLine label={resolvedLabel} className={className} />;
-    }
-    return (
-      <div
-        className={`flex min-w-0 items-center justify-between gap-3 ${className}`.trim()}
-      >
-        <LiveStatusLine label={resolvedLabel} className="min-w-0" />
-        <div className="shrink-0">{statusAccessory}</div>
-      </div>
-    );
-  };
-
-  const timeLabel = chatMessageTimeLabel(createdAt);
-
-  return (
-    <div
-      className="group/assistant-turn relative flex min-w-0 flex-col items-start animate-in fade-in-0 slide-in-from-bottom-1 duration-200"
-    >
-      <div className="flex w-full min-w-0 items-end gap-2">
-        <div className="w-5 shrink-0">
-          {showAvatar && workspaceId ? (
-            <AgentAvatar seed={workspaceId} size="sm" />
-          ) : null}
-        </div>
-      <article
-        className={
-          fitToContent
-            ? "min-w-0 inline-flex w-fit max-w-full flex-col rounded-lg bg-fg-6 px-3 py-2"
-            : "min-w-0 w-full max-w-4xl rounded-lg bg-fg-6 px-3 py-2"
-        }
-      >
-        {showStatusPlaceholder ? renderStatusLine(normalizedStatus) : null}
-
-        {renderedSegments.map((segment, index) =>
-          segment.kind === "execution" ? (
-            <TraceStepGroup
-              key={`execution-${index}`}
-              items={segment.items}
-              collapsedByStepId={collapsedTraceByStepId}
-              onToggleStep={onToggleTraceStep}
-              live={live}
-              liveOutputStarted={
-                live &&
-                renderedSegments
-                  .slice(index + 1)
-                  .some((nextSegment) => nextSegment.kind === "output")
-              }
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-              forceExpandToken={forceExpandToken}
-            />
-          ) : segment.tone === "error" ? (
-            <div
-              key={`output-${index}`}
-              className="theme-chat-system-bubble mt-2 first:mt-0 rounded-xl border px-3 py-2.5 text-xs text-foreground"
-            >
-              <div className="flex items-center gap-2">
-                <AlertTriangle className="size-3.5 shrink-0 text-destructive" />
-                <SimpleMarkdown
-                  className="chat-markdown max-w-full text-foreground"
-                  onLinkClick={onLinkClick}
-                  onLocalLinkClick={onLocalLinkClick}
-                >
-                  {segment.text}
-                </SimpleMarkdown>
-              </div>
-            </div>
-          ) : (
-            <SimpleMarkdown
-              key={`output-${index}`}
-              className="chat-markdown chat-assistant-markdown mt-2 first:mt-0 max-w-full text-foreground"
-              onLinkClick={onLinkClick}
-              onLocalLinkClick={onLocalLinkClick}
-            >
-              {segment.text}
-            </SimpleMarkdown>
-          ),
-        )}
-
-        {showWorkingStatusLine
-          ? renderStatusLine(
-              "Working",
-              renderedSegments.some((segment) => segment.kind === "execution")
-                ? ""
-                : "",
-            )
-          : null}
-
-        {showStreamingCursor ? <StreamingCursor /> : null}
-
-        {footerAccessory ? (
-          <div className="mt-2 flex justify-start">{footerAccessory}</div>
-        ) : null}
-
-        {memoryProposals.length > 0 ? (
-          <AssistantTurnMemoryProposals
-            proposals={memoryProposals}
-            proposalAction={memoryProposalAction}
-            editingProposalId={editingMemoryProposalId}
-            drafts={memoryProposalDrafts}
-            onEditProposal={onEditMemoryProposal}
-            onDraftChange={onMemoryProposalDraftChange}
-            onAcceptProposal={onAcceptMemoryProposal}
-            onDismissProposal={onDismissMemoryProposal}
-          />
-        ) : null}
-
-        {outputs.length > 0 ? (
-          <AssistantTurnOutputs
-            outputs={outputs}
-            onOpenOutput={onOpenOutput}
-            onOpenAllArtifacts={onOpenAllArtifacts}
-          />
-        ) : null}
-      </article>
-
-      </div>
-      {showActionsMenu || (showAvatar && timeLabel) ? (
-        <div className="flex h-6 items-center gap-2 pl-9">
-          {showAvatar && timeLabel ? (
-            <span className="select-none text-[10px] leading-none text-muted-foreground tabular-nums">
-              {timeLabel}
-            </span>
-          ) : null}
-          {showActionsMenu ? (
-            <div className="pointer-events-none opacity-0 transition-opacity duration-150 group-hover/assistant-turn:pointer-events-auto group-hover/assistant-turn:opacity-100 group-focus-within/assistant-turn:pointer-events-auto group-focus-within/assistant-turn:opacity-100">
-              <AssistantTurnActionsMenu
-                copyText={copyText}
-                hasFileEdits={hasFileEdits}
-                onViewFileChanges={
-                  hasFileEdits
-                    ? () => setForceExpandToken((token) => token + 1)
-                    : undefined
-                }
-                onViewTurnDetails={
-                  executionItems.length > 0
-                    ? () => setForceExpandToken((token) => token + 1)
-                    : undefined
-                }
-              />
-            </div>
-          ) : null}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function AssistantTurnOutputs({
-  outputs,
-  onOpenOutput,
-  onOpenAllArtifacts,
-}: {
-  outputs: WorkspaceOutputRecordPayload[];
-  onOpenOutput?: (output: WorkspaceOutputRecordPayload) => void;
-  onOpenAllArtifacts: (outputs: WorkspaceOutputRecordPayload[]) => void;
-}) {
-  const displayOutputs =
-    outputs.length > 1 ? dedupeOutputsForDisplay(outputs) : outputs;
-  return (
-    <div className="mt-3 flex flex-col gap-2">
-      {displayOutputs.map((output) => (
-        <button
-          key={output.id}
-          type="button"
-          onClick={() => onOpenOutput?.(output)}
-          className="group flex max-w-[380px] items-center gap-3 rounded-xl border border-border bg-card px-3 py-2.5 text-left transition-colors hover:border-border hover:bg-accent/50 disabled:cursor-default disabled:hover:border-border disabled:hover:bg-card"
-          disabled={!onOpenOutput}
-        >
-          <OutputArtifactIcon output={output} />
-          <div className="min-w-0 flex-1">
-            <div className="truncate text-sm font-medium text-foreground">
-              {output.title || "Untitled artifact"}
-            </div>
-            <div className="truncate text-xs text-muted-foreground">
-              {outputSecondaryLabel(output)}
-            </div>
-          </div>
-          <ArrowUpRight className="size-3.5 shrink-0 text-muted-foreground/0 transition-colors group-hover:text-muted-foreground" />
-        </button>
-      ))}
-
-      {displayOutputs.length > 1 ? (
-        <button
-          type="button"
-          onClick={() => onOpenAllArtifacts(displayOutputs)}
-          className="flex max-w-[380px] items-center gap-3 rounded-xl border border-dashed border-border px-3 py-2 text-left text-muted-foreground transition-colors hover:border-border hover:bg-accent/40 hover:text-foreground"
-        >
-          <div className="grid size-7 shrink-0 place-items-center rounded-md bg-muted text-muted-foreground">
-            <Folder className="size-3.5" />
-          </div>
-          <span className="text-xs">
-            View artifacts in this reply ({displayOutputs.length})
-          </span>
-        </button>
-      ) : null}
-    </div>
-  );
-}
-
-function memoryProposalStateLabel(state: MemoryUpdateProposalState) {
-  switch (state) {
-    case "accepted":
-      return "Saved";
-    case "dismissed":
-      return "Dismissed";
-    default:
-      return "Review";
-  }
-}
-
-function AssistantTurnMemoryProposals({
-  proposals,
-  proposalAction,
-  editingProposalId,
-  drafts,
-  onEditProposal,
-  onDraftChange,
-  onAcceptProposal,
-  onDismissProposal,
-}: {
-  proposals: MemoryUpdateProposalRecordPayload[];
-  proposalAction: { proposalId: string; action: "accept" | "dismiss" } | null;
-  editingProposalId: string | null;
-  drafts: Record<string, string>;
-  onEditProposal: (proposalId: string) => void;
-  onDraftChange: (proposalId: string, value: string) => void;
-  onAcceptProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-  onDismissProposal: (proposal: MemoryUpdateProposalRecordPayload) => void;
-}) {
-  return (
-    <div className="mt-4 grid gap-3">
-      {proposals.map((proposal) => {
-        const isPending = proposal.state === "pending";
-        const isEditing = editingProposalId === proposal.proposal_id;
-        const isActing = proposalAction?.proposalId === proposal.proposal_id;
-        const draftValue = drafts[proposal.proposal_id] ?? proposal.summary;
-
-        return (
-          <article
-            key={proposal.proposal_id}
-            className="bg-card rounded-2xl border border-border px-4 py-4 shadow-sm"
-          >
-            <div className="flex items-start justify-between gap-3">
-              <div className="min-w-0 flex-1">
-                <div className="flex items-center gap-2 text-sm font-medium text-foreground">
-                  <Lightbulb className="size-4 shrink-0 text-primary" />
-                  <span>{proposal.title}</span>
-                </div>
-                {isEditing ? (
-                  <textarea
-                    value={draftValue}
-                    onChange={(event) =>
-                      onDraftChange(proposal.proposal_id, event.target.value)
-                    }
-                    className="bg-muted mt-3 min-h-[86px] w-full rounded-xl border border-border px-3 py-2 text-sm leading-6 text-foreground outline-none transition focus:border-primary"
-                  />
-                ) : (
-                  <div className="mt-3 text-sm leading-6 text-muted-foreground">
-                    {proposal.summary}
-                  </div>
-                )}
-                {proposal.evidence ? (
-                  <div className="mt-3 text-xs leading-5 text-muted-foreground">
-                    {proposal.evidence}
-                  </div>
-                ) : null}
-              </div>
-
-              <div className="flex shrink-0 items-start gap-2">
-                <Badge variant="outline" className="uppercase">
-                  {memoryProposalStateLabel(proposal.state)}
-                </Badge>
-                {isPending ? (
-                  <Button
-                    type="button"
-                    variant="outline"
-                    size="icon"
-                    onClick={() => onEditProposal(proposal.proposal_id)}
-                    className="rounded-xl"
-                    aria-label="Edit memory proposal"
-                  >
-                    <PencilLine className="size-3.5" />
-                  </Button>
-                ) : null}
-              </div>
-            </div>
-
-            {isPending ? (
-              <div className="mt-4 flex flex-wrap gap-2">
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="lg"
-                  onClick={() => onDismissProposal(proposal)}
-                  disabled={isActing}
-                  className="rounded-2xl"
-                >
-                  {isActing && proposalAction?.action === "dismiss" ? (
-                    <Loader2 className="size-3 animate-spin" />
-                  ) : (
-                    <X className="size-3" />
-                  )}
-                  <span>Dismiss</span>
-                </Button>
-                <Button
-                  type="button"
-                  variant="outline"
-                  size="lg"
-                  onClick={() => onAcceptProposal(proposal)}
-                  disabled={isActing}
-                  className="rounded-2xl border-primary bg-primary/10 text-primary hover:bg-primary/14"
-                >
-                  {isActing && proposalAction?.action === "accept" ? (
-                    <Loader2 className="size-3 animate-spin" />
-                  ) : (
-                    <Check className="size-3" />
-                  )}
-                  <span>Accept</span>
-                </Button>
-              </div>
-            ) : null}
-          </article>
-        );
-      })}
-    </div>
-  );
-}
-
 export function ConversationTurns<Message extends ChatMessage>({
   messages,
   assistantLabel,
@@ -8892,184 +8007,6 @@ export function ConversationTurns<Message extends ChatMessage>({
         />
       ) : null}
     </>
-  );
-}
-
-function traceStatusLabel(status: ChatTraceStepStatus) {
-  if (status === "completed") {
-    return "Completed";
-  }
-  if (status === "error") {
-    return "Error";
-  }
-  if (status === "waiting") {
-    return "Waiting";
-  }
-  return "In progress";
-}
-
-function summarizeThinking(text: string) {
-  const firstContentLine =
-    text
-      .split("\n")
-      .map((line) => line.replace(/[*_`#>-]/g, "").trim())
-      .find(Boolean) || "Reasoning available";
-
-  return firstContentLine.length > 88
-    ? `${firstContentLine.slice(0, 85).trimEnd()}...`
-    : firstContentLine;
-}
-
-function TraceStepGroup({
-  items,
-  collapsedByStepId,
-  onToggleStep,
-  live = false,
-  liveOutputStarted = false,
-  onLinkClick,
-  onLocalLinkClick,
-  forceExpandToken = 0,
-}: {
-  items: ChatExecutionTimelineItem[];
-  collapsedByStepId: Record<string, boolean>;
-  onToggleStep: (stepId: string) => void;
-  live?: boolean;
-  liveOutputStarted?: boolean;
-  onLinkClick?: (url: string) => void;
-  onLocalLinkClick?: (href: string) => void;
-  forceExpandToken?: number;
-}) {
-  const steps = traceStepsFromExecutionItems(items);
-  const [groupExpanded, setGroupExpanded] = useState(
-    live && !liveOutputStarted,
-  );
-  const previousLiveRef = useRef(live);
-  const previousLiveOutputStartedRef = useRef(liveOutputStarted);
-  const previousForceExpandTokenRef = useRef(forceExpandToken);
-
-  useEffect(() => {
-    if (forceExpandToken !== previousForceExpandTokenRef.current) {
-      previousForceExpandTokenRef.current = forceExpandToken;
-      if (forceExpandToken > 0) {
-        setGroupExpanded(true);
-      }
-    }
-  }, [forceExpandToken]);
-
-  useEffect(() => {
-    if (live && !previousLiveRef.current) {
-      setGroupExpanded(!liveOutputStarted);
-    }
-    if (live && liveOutputStarted && !previousLiveOutputStartedRef.current) {
-      setGroupExpanded(false);
-    }
-    previousLiveRef.current = live;
-    previousLiveOutputStartedRef.current = liveOutputStarted;
-  }, [live, liveOutputStarted]);
-  const runningCount = steps.filter((s) => s.status === "running").length;
-  const terminalErrorCount = steps.filter(
-    (step) => step.kind === "phase" && step.status === "error",
-  ).length;
-  const groupHasTerminalError = terminalErrorCount > 0;
-  const stepCount = steps.length;
-  const stepLabel = `${stepCount} step${stepCount === 1 ? "" : "s"}`;
-  const activeStep =
-    [...steps]
-      .reverse()
-      .find((step) => step.status === "running" || step.status === "waiting") ??
-    null;
-  const groupIsLive = live && activeStep !== null && !groupHasTerminalError;
-  const latestStep = steps.length > 0 ? steps[steps.length - 1] : null;
-  const latestThinkingItem =
-    [...items]
-      .reverse()
-      .find(
-        (
-          item,
-        ): item is Extract<ChatExecutionTimelineItem, { kind: "thinking" }> =>
-          item.kind === "thinking",
-      ) ?? null;
-  const summaryStep = activeStep ?? (groupIsLive ? latestStep : null);
-  const summarySuffix = groupHasTerminalError
-    ? ` (${terminalErrorCount} failed)`
-    : "";
-  const showLiveSummarySpinner =
-    (groupIsLive || runningCount > 0) && !groupExpanded;
-  const summaryLabel = summaryStep
-    ? summaryStep === activeStep || summaryStep.status === "waiting"
-      ? `${traceStatusLabel(summaryStep.status)}: ${summaryStep.title}`
-      : groupIsLive
-        ? summaryStep.title
-        : `${traceStatusLabel(summaryStep.status)}: ${summaryStep.title}`
-    : groupIsLive
-      ? latestThinkingItem
-        ? summarizeThinking(latestThinkingItem.text)
-        : stepCount > 0
-          ? `Working through ${stepLabel}...`
-          : "Thinking..."
-      : runningCount > 0
-        ? `Running ${stepLabel}...`
-        : latestThinkingItem
-          ? summarizeThinking(latestThinkingItem.text)
-          : stepCount > 0 && latestStep
-            ? latestStep.title
-            : "Execution trace";
-
-  return (
-    <div className="mt-3 first:mt-0">
-      <button
-        type="button"
-        onClick={() => setGroupExpanded((v) => !v)}
-        className="flex w-full items-center gap-2 rounded-lg px-2.5 py-1.5 -ml-2.5 text-left text-xs text-muted-foreground transition-colors hover:bg-muted"
-      >
-        {groupHasTerminalError ? (
-          <AlertTriangle className="size-3.5 shrink-0 text-destructive" />
-        ) : showLiveSummarySpinner ? (
-          <Loader2 className="size-3.5 shrink-0 animate-spin text-muted-foreground" />
-        ) : groupIsLive || runningCount > 0 ? (
-          <Clock3 className="size-3.5 shrink-0 text-muted-foreground" />
-        ) : (
-          <Check className="size-3.5 shrink-0 text-success" />
-        )}
-        <span className="min-w-0 flex-1 truncate leading-5">
-          {summaryLabel}
-          {summarySuffix}
-        </span>
-        {stepCount > 0 && !groupIsLive && !groupHasTerminalError ? (
-          <span
-            aria-hidden
-            className="shrink-0 rounded-full bg-muted px-1.5 py-px text-[10px] tabular-nums text-muted-foreground"
-          >
-            {stepCount}
-          </span>
-        ) : null}
-        <ChevronDown
-          className={`size-3 shrink-0 transition-transform ${groupExpanded ? "rotate-180" : ""}`}
-        />
-      </button>
-
-      {groupExpanded ? (
-        <div className="mt-1 ml-1 space-y-0.5">
-          {items.map((item) =>
-            item.kind === "thinking" ? (
-              <ExecutionTimelineThinkingEntry
-                key={item.id}
-                text={item.text}
-                onLinkClick={onLinkClick}
-                onLocalLinkClick={onLocalLinkClick}
-              />
-            ) : (
-              <TraceTimelineStepEntry
-                key={item.id}
-                step={item.step}
-                collapsedByStepId={collapsedByStepId}
-                onToggleStep={onToggleStep}
-              />
-            ),
-          )}
-        </div>
-      ) : null}
-    </div>
   );
 }
